Avoid view unnecessarily retained while a notification is being displayed
The SRF blocked segment example starting with a skip notification was not letting playback stop correctly if dismissed while the notification was still displayed. The reason was that a perform selector with delay retains objects, and unregistration was not made correctly (the selector was not the one for which registration had been initially made, and proper cleanup was missing).

Since the view retains the controller, playback was therefore not stoping immediately, but after 5 seconds (the scheduled duration for the perform selector).